Renombrado el antiguo MitoTextField y añadido uno nuevo, con el esqueleto de lo que se debería montar.
